🐛 修复 RSS 列表项 `key` 类型不匹配导致的潜在问题

+++ common-client/src/anime/pages/Rss.kt
@@ -134,7 +134,7 @@ fun RssList(
             }
         }
         LazyColumn(modifier = Modifier.fillMaxSize(), lazyListState = rssListState) {
-            items(rssList.value, key = { it.id }) { rss ->
+            items(rssList.value, key = { it.id.toLong() }) { rss ->
                 Card(
                     modifier = Modifier.fillMaxWidth(),
                     onClick = {
